% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_z_test.R
\name{corrected_z_test}
\alias{corrected_z_test}
\title{corrected_z_test}
\usage{
corrected_z_test(data, sig_level = 0.05)
}
\arguments{
\item{data}{partially paired data}

\item{sig_level}{significance level for the hypothesis test}
}
\value{
list object containing in particular:
z_corr, the corrected z statistic for the hypothesis test
stat_sig, the result of the hypothesis test
}
\description{
corrected_z_test
}
\examples{
corrected_z_test(data = simulate_pp_data())
}
