% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pp_data.R
\name{simulate_pp_data}
\alias{simulate_pp_data}
\title{simulate_pp_data}
\usage{
simulate_pp_data(
  n = 4,
  n0 = 4,
  n1 = 4,
  delta = 0.5,
  sigma_0 = 1,
  sigma_1 = 2,
  rho = 0.5,
  tidy = TRUE,
  seed = 1234
)
}
\arguments{
\item{n}{number of paired observations}

\item{n0}{number of unpaired observations in the pre-group}

\item{n1}{number of unpaired observations in the post-group}

\item{delta}{effect size}

\item{sigma_0}{standard deviation of pre-treatment measurements}

\item{sigma_1}{standard deviation of post-treatment measurements}

\item{rho}{correlation between pre- and post- treatment measurements}

\item{tidy}{whether to return the simulated data in tidy format}

\item{seed}{seed passed to set.seed prior to simulation}
}
\value{

}
\description{
simulate_pp_data
}
